requirements.txt
matplotlib
Pillow
torch
torchvision
transformers
tqdm
